Move test_data_kw_define to test_enkf_main
